+++ src/core/geo/frustum.py
@@ -36,7 +36,7 @@ class Frustum:
 
     def __init__(self, fovy: float = 60.0, aspect_ratio: float = 1.0,
                  orthogonal: bool = False,
-                 orthogonal_size: tuple[int, int] = (16, 16),
+                 orthogonal_size: tuple[float, float] = (16, 16),
                  near: float = 0.1, far: float = 10000,
                  transform: Optional[Transform] = None
                  ):
